Phase 1: 16-bit tiles, tile metadata schema, protocol version handshake
Chunk fg/bg widened to Uint16Array (removes the 256-tile ceiling; liquid stays
8-bit). TileProps gains optional structured metadata (tier/mergeGroup/wall/
blastResistance/flags/toolType/soundGroup) exposed via category-defaulting helper
functions so adding tiles never touches every row; tile ids documented as
stable/append-only. Net: added PROTOCOL_VERSION to the hello/welcome handshake;
mismatched versions are refused (host closes the conn, client errors out).
Verified headlessly (scripts/phase1-check.ts): determinism intact, a >255 id
places, replicates via deltas, and survives unload/reload.

+++ src/world/Chunk.ts
@@ -18,15 +18,17 @@ export function tileIndex(lx: number, ly: number): number {
 export class Chunk {
   readonly cx: number;
   readonly cy: number;
-  readonly fg: Uint8Array; // foreground tile ids
-  readonly bg: Uint8Array; // background wall tile ids
+  // Tile ids are 16-bit so the world can hold hundreds of blocks/walls/furniture/ores/machines
+  // (well past the old 256 ceiling). Liquid stays 8-bit — it encodes a level + type flag, not an id.
+  readonly fg: Uint16Array; // foreground tile ids
+  readonly bg: Uint16Array; // background wall tile ids
   readonly liquid: Uint8Array; // liquid state (see Liquid.ts encoding)
 
   constructor(cx: number, cy: number) {
     this.cx = cx;
     this.cy = cy;
-    this.fg = new Uint8Array(AREA); // defaults to Air (0)
-    this.bg = new Uint8Array(AREA);
+    this.fg = new Uint16Array(AREA); // defaults to Air (0)
+    this.bg = new Uint16Array(AREA);
     this.liquid = new Uint8Array(AREA);
   }
